extraced render heat layer from main file.
